refactor: move up comment modules
