+++ backend/apps/api/urls.py
@@ -1,0 +1,8 @@
+from rest_framework import routers
+
+from .viewsets.comment import CommentViewSet
+
+router = routers.DefaultRouter()
+router.register(r'comments', CommentViewSet)
+
+urlpatterns = router.urls
