--- backend/apps/comment/api/v1/urls.py
@@ -1,8 +1,0 @@
-from rest_framework import routers
-
-from .viewsets import CommentViewSet
-
-router = routers.DefaultRouter()
-router.register(r'', CommentViewSet)
-
-urlpatterns = router.urls
